# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.AddMemberJSON:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      role_id:
        type: string
    required:
    - email
    - name
    - password
    type: object
  controllers.ChangePwdJSON:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  controllers.RegisterJSON:
    properties:
      email:
        type: string
      invite_code:
        description: "\U0001F525"
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - invite_code
    - name
    - password
    type: object
...
  controllers.UpdateProfileJSON:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
  controllers.UpdateUserJSON:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role_id:
        type: string
    type: object
...
  models.Family:
    properties:
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.Role:
    properties:
      can_edit_schema:
        type: boolean
      can_manage_users:
        type: boolean
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      description:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      base_currency:
        description: Налаштування користувача
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      email:
        description: Додав unique
        type: string
      family:
        $ref: '#/definitions/models.Family'
      family_id:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      language:
        type: string
      name:
        type: string
      password_hash:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: Relations
      role_id:
        type: string
      theme:
        type: string
      updated_at:
        type: integer
    type: object
...
  services.LoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
...
info:
  contact:
    name: Vlad Hrytsaiuk
  description: API for WeGaS Finance application.
  title: WeGaS Finance API
  version: "1.0"
paths:
...
  /family/users:
    post:
      consumes:
      - application/json
      description: Adds a new user to the current user's family. Only accessible by
        family parents/admins.
      parameters:
      - description: Member details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.AddMemberJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a family member
      tags:
      - Users
...
  /users:
    get:
      consumes:
      - application/json
      description: Returns a list of all members in the current user's family.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get family members
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Register a new user and create a family for them. Requires a valid
        invite code.
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.RegisterJSON'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/services.LoginResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Invalid invite code
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a family member. Only accessible by family parents/admins.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a family member
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates information for a specific family member. Only accessible
        by family parents/admins.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated member details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateUserJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a family member
      tags:
      - Users
  /users/me:
    get:
      consumes:
      - application/json
      description: Returns the profile of the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get current user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates the profile information for the currently authenticated
        user.
      parameters:
      - description: Updated profile details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateProfileJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update current user profile
      tags:
      - Users
  /users/password:
    put:
      consumes:
      - application/json
      description: Changes the password for the currently authenticated user.
      parameters:
      - description: Password details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.ChangePwdJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or old password
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Change current user password
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"